🔨Move App component and related files to its own folder
