+++ src/index.js
@@ -1,7 +1,7 @@
 import React from "react"
 import { createRoot } from "react-dom"
 import "./index.css"
-import App from "./App"
+import App from "./components/App/App"
 import reportWebVitals from "./reportWebVitals"
 
 const container = document.getElementById("root")
